+++ src/Products/MeetingSeraing/adapters.py
@@ -468,9 +468,6 @@ class CustomSeraingMeeting(CustomMeeting):
         return res
 
 
-old_setTakenOverBy = MeetingItem.setTakenOverBy
-
-
 class CustomSeraingMeetingItem(CustomMeetingItem):
     """Adapter that adapts a meeting item implementing IMeetingItem to the
        interface IMeetingItemCustom."""
@@ -529,20 +526,44 @@ class CustomSeraingMeetingItem(CustomMeetingItem):
 
     def setTakenOverBy(self, value, **kwargs):
         # call original method
-        old_setTakenOverBy(self, value, **kwargs)
-        item = self.getSelf()
-        if not item._at_creation_flag:
-            wf_states_to_keep = ['presented', 'validated_by_dg', 'itemfrozen', 'accepted_but_modified', 'accepted']
-            if item.queryState() in wf_states_to_keep:
-                tool = getToolByName(item, 'portal_plonemeeting')
-                cfg = tool.getMeetingConfig(item)
-                for wf_state_to_keep in wf_states_to_keep:
-                    wf_state = "%s__wfstate__%s" % (cfg.getItemWorkflow(), wf_state_to_keep)
-                    if value:
-                        item.takenOverByInfos[wf_state] = value
-                    elif not value and wf_state in item.takenOverByInfos:
-                        del item.takenOverByInfos[wf_state]
-        item.getField('takenOverBy').set(item, value, **kwargs)
+        if not self._at_creation_flag:
+            # save takenOverBy to takenOverByInfos for current review_state
+            # or check for a wf_state in kwargs
+            if 'wf_state' in kwargs:
+                wf_state = kwargs['wf_state']
+            else:
+                tool = api.portal.get_tool('portal_plonemeeting')
+                cfg = tool.getMeetingConfig(self)
+                wf_state = "%s__wfstate__%s" % (cfg.getItemWorkflow(), self.queryState())
+            if value:
+                self.takenOverByInfos[wf_state] = value
+                # xxx for Seraing, in some states, we would keep the "Taken over" for some next states
+                wf_states_service = ['itemcreated', 'proposed_to_servicehead', 'proposed_to_officemanager',
+                                     'proposed_to_divisionhead', 'proposed' ]
+                wf_state_gs = ['validated','presented', 'validated_by_dg', 'itemfrozen']
+                wf_state_close = ['accepted', 'accepted_but_closed', 'accepted_but_modified',
+                                  'accepted_but_modified_but_closed', 'delayed', 'delayed_closed']
+                wf_states = []
+                wf_states_to_use = []
+                cpt = 0
+                if self.queryState() in wf_states_service:
+                    wf_states_to_use = wf_states_service
+                elif self.queryState() in wf_state_gs:
+                    wf_states_to_use = wf_state_gs
+                elif self.queryState() in wf_state_close:
+                    wf_states_to_use
+                for wf_state_service in wf_states_to_use:
+                    if self.queryState() == wf_state_service:
+                        break
+                    cpt += 1
+                wf_states = wf_states_to_use[cpt:]
+                # add for next states the taken over info
+                for wf_state in wf_states:
+                    wf_state = "%s__wfstate__%s" % (cfg.getItemWorkflow(), wf_state)
+                    self.takenOverByInfos[wf_state] = value
+            elif not value and wf_state in self.takenOverByInfos:
+                del self.takenOverByInfos[wf_state]
+        self.getField('takenOverBy').set(self, value, **kwargs)
 
     MeetingItem.setTakenOverBy = setTakenOverBy